update process manager unit tests
