--- tests/unit/managers/test_process_manager.py
@@ -1,1437 +1,0 @@
-# SPDX-FileCopyrightText: 2021 Sebastian Garcia <sebastian.garcia@agents.fel.cvut.cz>
-# SPDX-License-Identifier: GPL-2.0-only
-import json
-import signal
-import ast
-import pytest
-from unittest.mock import Mock, call, patch
-from managers.process_manager.process_manager import ProcessManager
-from modules.supported_module_names import Modules
-from tests.module_factory import ModuleFactory
-from slips_files.common.slips_utils import utils
-from slips_files.common.input_type import InputType
-
-
-@pytest.mark.parametrize(
-    "input_type, input_information, cli_packet_filter, "
-    "zeek_or_bro, line_type",
-    [
-        # Test case 1: pcap input
-        (
-            InputType.PCAP,
-            "test.pcap",
-            "tcp port 80",
-            "zeek",
-            "conn",
-        ),
-        # Test case 2: zeek input
-        (InputType.ZEEK, "test.log", "", "bro", "dns"),
-        # Test case 3: stdin input
-        (InputType.STDIN, "-", "", "zeek", "http"),
-    ],
-)
-def test_start_input_process(
-    input_type,
-    input_information,
-    cli_packet_filter,
-    zeek_or_bro,
-    line_type,
-):
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.input_type = input_type
-    process_manager.main.input_information = input_information
-    process_manager.main.args.pcapfilter = cli_packet_filter
-    process_manager.main.zeek_bro = zeek_or_bro
-    process_manager.main.line_type = line_type
-    process_manager.main.bloom_filters_man = Mock()
-
-    with patch("managers.process_manager.startup_mixin.Input") as mock_input:
-        mock_input_process = Mock()
-        mock_input.return_value = mock_input_process
-        mock_input_process.pid = 54321
-
-        result = process_manager.start_input_process()
-
-        assert result == mock_input_process
-        mock_input.assert_called_once_with(
-            process_manager.main.logger,
-            process_manager.main.args.output,
-            process_manager.main.redis_port,
-            process_manager.termination_event,
-            process_manager.main.args,
-            process_manager.main.conf,
-            process_manager.main.pid,
-            process_manager.main.bloom_filters_man,
-            is_input_done=process_manager.is_input_done,
-            profiler_queue=process_manager.profiler_queue,
-            input_type=input_type,
-            input_information=input_information,
-            cli_packet_filter=cli_packet_filter,
-            zeek_or_bro=zeek_or_bro,
-            line_type=line_type,
-            is_profiler_done_event=process_manager.is_profiler_done_event,
-            is_input_done_event=process_manager.is_input_done_event,
-            is_input_failed_event=process_manager.is_input_failed_event,
-            is_slips_live_updating_event=(
-                process_manager.is_slips_live_updating_event
-            ),
-            is_profiler_done_starting_initial_workers_event=(
-                process_manager.is_profiler_done_starting_initial_workers_event
-            ),
-        )
-        mock_input_process.start.assert_called_once()
-        process_manager.main.print.assert_called_once()
-        process_manager.main.db.store_pid.assert_called_once_with(
-            "Input", 54321
-        )
-
-
-@pytest.mark.parametrize(
-    "is_slips_started_by_an_update,expected_is_set",
-    [(True, False), (False, False)],
-)
-def test_init_sets_live_update_event_for_update_start(
-    is_slips_started_by_an_update,
-    expected_is_set,
-):
-    """Test that ProcessManager starts with an unset live-update event."""
-    main_mock = Mock()
-    main_mock.args.is_slips_started_by_an_update = (
-        is_slips_started_by_an_update
-    )
-    main_mock.conf.read_configuration.side_effect = (
-        lambda section, name, default_value: default_value
-    )
-    main_mock.conf.get_bootstrapping_modules.return_value = []
-
-    process_manager = ProcessManager(main_mock)
-
-    assert process_manager.is_slips_live_updating_event.is_set() is (
-        expected_is_set
-    )
-
-
-@pytest.mark.parametrize(
-    "module_name, modules_to_ignore, expected",
-    [
-        # Test case 1: Module name not in ignore list
-        ("test_module", ["ignore_module"], False),
-        # Test case 2: Exact match in ignore list
-        ("ignore_module", ["ignore_module"], True),
-        # Test case 3: Partial match in ignore list
-        ("test_ignore_module", ["ignore_module"], True),
-        # Test case 4: Module name with spaces, not in ignore list
-        ("test module", ["ignore module"], False),
-        # Test case 5: Module name with hyphens, not in ignore list
-        ("test-module", ["ignore-module"], False),
-    ],
-)
-def test_is_ignored_module(module_name, modules_to_ignore, expected):
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.user_disabled_modules = set(modules_to_ignore)
-    process_manager.slips_disabled_modules = set()
-    assert process_manager.is_disabled_module(module_name) == expected
-
-
-def test_get_disabled_modules_uses_disabled_module_helpers() -> None:
-    """Test disabled modules are returned from the dedicated helper methods."""
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.get_user_disabled_modules = Mock(
-        return_value={Modules.TEMPLATE}
-    )
-    process_manager.get_runtime_disabled_modules = Mock(
-        return_value={Modules.BLOCKING}
-    )
-
-    disabled_modules = process_manager.get_disabled_modules()
-
-    assert disabled_modules == ({Modules.TEMPLATE}, {Modules.BLOCKING})
-    assert process_manager.user_disabled_modules == {Modules.TEMPLATE}
-    assert process_manager.slips_disabled_modules == {Modules.BLOCKING}
-    process_manager.get_user_disabled_modules.assert_called_once_with()
-    process_manager.get_runtime_disabled_modules.assert_called_once_with()
-
-
-@pytest.mark.parametrize(
-    "configured_modules, expected_modules",
-    [
-        ([" template ", "custom_module"], {Modules.TEMPLATE, "custom_module"}),
-        (
-            [" template ", " feeds_update_manager "],
-            {Modules.TEMPLATE, Modules.FEEDS_UPDATE_MANAGER},
-        ),
-        ([], set()),
-    ],
-)
-def test_get_user_disabled_modules(
-    configured_modules: list, expected_modules: set
-) -> None:
-    """Test user-disabled modules are read from config and stripped."""
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.conf.llm_enabled.return_value = True
-    process_manager.main.conf.alert_summary_enabled.return_value = True
-    process_manager.main.conf.regex_generator_enabled.return_value = True
-    process_manager.main.conf.read_configuration.reset_mock()
-    process_manager.main.conf.read_configuration.side_effect = None
-    process_manager.main.conf.read_configuration.return_value = (
-        configured_modules
-    )
-
-    disabled_modules = process_manager.get_user_disabled_modules()
-
-    assert disabled_modules == expected_modules
-    process_manager.main.conf.read_configuration.assert_called_once_with(
-        "modules", "disable", ["template"]
-    )
-
-
-def test_get_slips_disabled_modules_recomputes_runtime_disabled_modules() -> (
-    None
-):
-    """Test runtime-disabled modules are recomputed from current settings."""
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.args.clearblocking = False
-    process_manager.main.args.blocking = False
-    process_manager.slips_disabled_modules = {
-        Modules.BLOCKING,
-        Modules.ARP_POISONER,
-    }
-
-    disabled_modules = process_manager.get_runtime_disabled_modules()
-
-    assert Modules.BLOCKING in disabled_modules
-    assert Modules.ARP_POISONER in disabled_modules
-
-
-@pytest.mark.parametrize(
-    "input_type, argv, export_to, expected_user, expected_slips",
-    [
-        (
-            InputType.PCAP,
-            ["slips.py", "-f", "dataset/test.pcap"],
-            [],
-            {Modules.TEMPLATE, "custom_module"},
-            {
-                Modules.EXPORTING_ALERTS,
-                Modules.P2P_TRUST,
-                Modules.FIDES,
-                Modules.IRIS,
-                Modules.CESNET,
-                Modules.BLOCKING,
-                Modules.ARP_POISONER,
-                Modules.CYST,
-            },
-        ),
-        (
-            InputType.ZEEK,
-            ["slips.py", "-f", "dataset/conn.log"],
-            ["stix"],
-            {Modules.TEMPLATE, "custom_module"},
-            {
-                Modules.P2P_TRUST,
-                Modules.FIDES,
-                Modules.IRIS,
-                Modules.CESNET,
-                Modules.BLOCKING,
-                Modules.ARP_POISONER,
-                Modules.LEAK_DETECTOR,
-                Modules.CYST,
-            },
-        ),
-    ],
-)
-def test_get_disabled_modules(
-    input_type: InputType,
-    argv: list,
-    export_to: list,
-    expected_user: set,
-    expected_slips: set,
-) -> None:
-    """Test disabled modules are split by user and Slips runtime rules."""
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.conf.llm_enabled.return_value = True
-    process_manager.main.conf.alert_summary_enabled.return_value = True
-    process_manager.main.conf.regex_generator_enabled.return_value = True
-    process_manager.main.input_type = input_type
-    process_manager.main.args.clearblocking = False
-    process_manager.main.args.blocking = False
-    process_manager.main.conf.read_configuration.side_effect = (
-        lambda section, name, default_value: (
-            [" template ", "custom_module"]
-            if (section, name) == ("modules", "disable")
-            else default_value
-        )
-    )
-    process_manager.main.conf.export_to.return_value = export_to
-
-    user_disabled_modules, slips_disabled_modules = (
-        process_manager.get_disabled_modules()
-    )
-
-    assert user_disabled_modules == expected_user
-    assert slips_disabled_modules == expected_slips
-
-
-def test_get_hitlist_in_order_uses_supported_module_name_values() -> None:
-    """Test hitlist lookups use the shared supported module names enum."""
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.args.blocking = True
-    process_manager.main.db.get_disabled_modules.return_value = []
-    process_manager.main.db.get_pid_of.side_effect = lambda module_name: {
-        "evidence_handler": 10,
-        Modules.BLOCKING: 20,
-        Modules.ARP_POISONER: 30,
-        Modules.EXPORTING_ALERTS: 40,
-    }.get(module_name)
-    process_manager.children = [
-        Mock(pid=10),
-        Mock(pid=20),
-        Mock(pid=30),
-        Mock(pid=40),
-        Mock(pid=50),
-    ]
-
-    to_kill_first, to_kill_last = process_manager.get_hitlist_in_order()
-
-    assert [process.pid for process in to_kill_first] == [50]
-    assert [process.pid for process in to_kill_last] == [10, 20, 30, 40]
-
-
-def test_get_user_disabled_modules_strips_configured_entries() -> None:
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.conf.llm_enabled.return_value = True
-    process_manager.main.conf.alert_summary_enabled.return_value = True
-    process_manager.main.conf.regex_generator_enabled.return_value = True
-    process_manager.main.conf.read_configuration.side_effect = None
-    process_manager.main.conf.read_configuration.return_value = [
-        " template ",
-        " custom_module",
-    ]
-
-    assert process_manager.get_user_disabled_modules() == {
-        Modules.TEMPLATE,
-        "custom_module",
-    }
-
-
-@pytest.mark.parametrize(
-    "llm_enabled, alert_summary_enabled, regex_generator_enabled, expected",
-    [
-        (False, True, True, {Modules.LLM_PROXY}),
-        (True, False, True, {Modules.ALERT_SUMMARY}),
-        (True, True, False, {Modules.REGEX_GENERATOR}),
-        (
-            False,
-            False,
-            False,
-            {
-                Modules.LLM_PROXY,
-                Modules.ALERT_SUMMARY,
-                Modules.REGEX_GENERATOR,
-            },
-        ),
-    ],
-)
-def test_get_user_disabled_modules_includes_feature_toggled_modules(
-    llm_enabled: bool,
-    alert_summary_enabled: bool,
-    regex_generator_enabled: bool,
-    expected: set[Modules],
-) -> None:
-    """Test dedicated enabled flags disable LLM-related modules at startup."""
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.conf.read_configuration.side_effect = None
-    process_manager.main.conf.read_configuration.return_value = []
-    process_manager.main.conf.llm_enabled.return_value = llm_enabled
-    process_manager.main.conf.alert_summary_enabled.return_value = (
-        alert_summary_enabled
-    )
-    process_manager.main.conf.regex_generator_enabled.return_value = (
-        regex_generator_enabled
-    )
-
-    assert process_manager.get_user_disabled_modules() == expected
-
-
-def test_get_runtime_disabled_modules_uses_runtime_rules() -> None:
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.db.is_running_non_stop.return_value = False
-    process_manager.main.conf.export_to.return_value = []
-    process_manager.main.conf.use_local_p2p.return_value = False
-    process_manager.main.conf.use_global_p2p.return_value = False
-    process_manager.main.conf.send_to_warden.return_value = False
-    process_manager.main.conf.receive_from_warden.return_value = False
-    process_manager.main.args.clearblocking = False
-    process_manager.main.args.blocking = False
-    process_manager.main.input_type = InputType.ZEEK
-    process_manager.main.args.input_module = ""
-    process_manager.slips_disabled_modules = set()
-
-    assert process_manager.get_runtime_disabled_modules() == {
-        Modules.EXPORTING_ALERTS,
-        Modules.P2P_TRUST,
-        Modules.FIDES,
-        Modules.IRIS,
-        Modules.CESNET,
-        Modules.BLOCKING,
-        Modules.ARP_POISONER,
-        Modules.LEAK_DETECTOR,
-        Modules.CYST,
-    }
-
-
-def test_get_disabled_modules_keeps_unrelated_modules_enabled() -> None:
-    """Test disabling regex_generator does not disable unrelated modules."""
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.conf.read_configuration.side_effect = (
-        lambda section, name, default_value: (
-            [Modules.REGEX_GENERATOR]
-            if (section, name) == ("modules", "disable")
-            else default_value
-        )
-    )
-
-    _, slips_disabled_modules = process_manager.get_disabled_modules()
-
-    assert Modules.T_CELL not in slips_disabled_modules
-    assert Modules.T_CELL not in process_manager.slips_disabled_modules
-
-
-@pytest.mark.parametrize(
-    "disabled_modules, enabled_value, expected",
-    [
-        ({Modules.LLM_PROXY}, True, True),
-        (set(), False, False),
-        (set(), True, False),
-    ],
-)
-def test_has_missing_dependency_respects_disabled_and_configured_dependencies(
-    disabled_modules: set[Modules],
-    enabled_value: bool,
-    expected: bool,
-) -> None:
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.conf.read_configuration.side_effect = (
-        lambda section, name, default_value: (
-            enabled_value
-            if section == "llmproxy" and name == "enabled"
-            else default_value
-        )
-    )
-
-    assert (
-        process_manager._has_missing_dependency(
-            Modules.REGEX_GENERATOR, disabled_modules
-        )
-        is expected
-    )
-
-
-def test_read_config_parses_module_dependencies() -> None:
-    module_factory = ModuleFactory()
-    process_manager = module_factory.create_process_manager_obj()
-
-    assert process_manager.module_dependencies == {
-        Modules.LLM_PROXY: (),
-        Modules.REGEX_GENERATOR: (Modules.LLM_PROXY,),
-        Modules.ALERT_SUMMARY: (Modules.LLM_PROXY,),
-    }
-
-
-@pytest.mark.parametrize(
-    "module_name, expected_dependencies",
-    [
-        (Modules.LLM_PROXY, ()),
-        ("regex_generator", (Modules.LLM_PROXY,)),
-        (
-            "modules.alert_summary.alert_summary",
-            (Modules.LLM_PROXY,),
-        ),
-        ("unknown_module", ()),
-    ],
-)
-def test_get_module_dependencies(
-    module_name: str, expected_dependencies: tuple
-) -> None:
-    module_factory = ModuleFactory()
-    process_manager = module_factory.create_process_manager_obj()
-
-    assert (
-        process_manager.get_module_dependencies(module_name)
-        == expected_dependencies
-    )
-
-
-def test_get_dependency_disabled_modules_disables_llm_dependents() -> None:
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.user_disabled_modules = {Modules.LLM_PROXY}
-    process_manager.main.conf.read_configuration.side_effect = (
-        lambda section, name, default_value: default_value
-    )
-
-    assert process_manager._get_dependency_disabled_modules(set()) == {
-        Modules.REGEX_GENERATOR,
-        Modules.ALERT_SUMMARY,
-    }
-
-
-def test_get_dependency_disabled_modules_ignores_disabled_dependents() -> None:
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.user_disabled_modules = {
-        Modules.LLM_PROXY,
-        Modules.REGEX_GENERATOR,
-        Modules.ALERT_SUMMARY,
-    }
-    process_manager.main.conf.read_configuration.side_effect = (
-        lambda section, name, default_value: default_value
-    )
-
-    with patch.object(process_manager.main, "print") as mock_print:
-        assert process_manager._get_dependency_disabled_modules(set()) == set()
-
-    mock_print.assert_not_called()
-
-
-@pytest.mark.parametrize(
-    "bootstrapping_node, use_global_p2p, expected",
-    [
-        (True, True, True),
-        (True, False, False),
-        (False, True, False),
-    ],
-)
-def test_should_bootstrap_p2p(bootstrapping_node, use_global_p2p, expected):
-    """Test P2P bootstrapping is enabled only when both flags are set."""
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.bootstrapping_node = bootstrapping_node
-    process_manager.use_global_p2p = use_global_p2p
-    process_manager.main.db.is_running_non_stop.return_value = True
-
-    assert process_manager.is_bootstrapping_node() is expected
-
-
-@pytest.mark.parametrize(
-    "should_bootstrap, module_name, disabled_modules, expected",
-    [
-        (True, "modules.fides.fides", [], True),
-        (True, "modules.flowalerts.flowalerts", [], False),
-        (False, "modules.flowalerts.flowalerts", ["flowalerts"], False),
-        (False, "modules.fides.fides", [], True),
-    ],
-)
-def test_should_load_module(
-    should_bootstrap, module_name, disabled_modules, expected
-):
-    """Test module loading respects bootstrapping and disabled modules."""
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.bootstrapping_modules = ["fides", "iris"]
-    process_manager.user_disabled_modules = set(disabled_modules)
-    process_manager.slips_disabled_modules = set()
-    process_manager.is_bootstrapping_node = Mock(return_value=should_bootstrap)
-
-    assert process_manager._should_load_module(module_name) is expected
-
-
-def test_print_disabled_modules():
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.user_disabled_modules = {Modules.TEMPLATE, Modules.FIDES}
-    process_manager.slips_disabled_modules = set()
-    with patch.object(process_manager.main, "print") as mock_print:
-        process_manager.print_disabled_modules()
-        printed_modules = ast.literal_eval(
-            mock_print.call_args.args[0].removeprefix("Disabled Modules: ")
-        )
-
-        assert set(printed_modules) == {"template", "fides"}
-        mock_print.assert_called_once()
-
-
-@pytest.mark.parametrize(
-    "pending_module_names, expected_print_calls",
-    [
-        # Test case 1: No pending modules, no additional print calls
-        ([], 1),
-        # Test case 2: Pending modules without feeds_update_manager, one additional print call
-        (["Module1", "Module2"], 1),
-        # Test case 3: Pending update manager prints the extra warning
-        ([Modules.FEEDS_UPDATE_MANAGER], 2),
-    ],
-)
-def test_warn_about_pending_modules(
-    pending_module_names, expected_print_calls
-):
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.warning_printed_once = False
-    pending_modules = []
-    for module_name in pending_module_names:
-        pending_module = Mock()
-        pending_module.name = module_name
-        pending_modules.append(pending_module)
-
-    with patch.object(process_manager.main, "print") as mock_print:
-        process_manager.warn_about_pending_modules(pending_modules)
-
-        assert mock_print.call_count == expected_print_calls
-        expected_result = True
-        assert process_manager.warning_printed_once == expected_result
-
-    with patch.object(process_manager.main, "print") as mock_print:
-        process_manager.warn_about_pending_modules(pending_modules)
-
-        mock_print.assert_not_called()
-
-
-def test_reading_flows_from_cyst_uses_supported_module_name() -> None:
-    """Test CYST input detection relies on the shared module-name enum."""
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.args.input_module = Modules.CYST.value
-
-    assert process_manager._reading_flows_from_cyst() is True
-
-
-@pytest.mark.parametrize(
-    "blocking_enabled, exporting_alerts_disabled, "
-    "expected_kill_first, expected_kill_last",
-    [  # Testcase1: blocking enabled, exporting_alerts enabled
-        (True, False, [1, 2, 6, 7], [3, 4, 5]),
-        # Testcase2: blocking disabled, exporting_alerts enabled
-        (False, False, [1, 2, 4, 6, 7], [3, 5]),
-        # Testcase3: blocking enabled, exporting_alerts disabled
-        (True, True, [1, 2, 5, 6, 7], [3, 4]),
-        # Testcase4: blocking disabled, exporting_alerts disabled
-        (False, True, [1, 2, 4, 5, 6, 7], [3]),
-        # Testcase5: All enabled, some PIDs are None
-        (True, False, [1, 2, 6, 7], [3, 4, 5]),
-    ],
-)
-def test_get_hitlist_in_order(
-    blocking_enabled,
-    exporting_alerts_disabled,
-    expected_kill_first,
-    expected_kill_last,
-):
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.children = [
-        Mock(pid=1, name="Process1"),
-        Mock(pid=2, name="Process2"),
-        Mock(pid=3, name="evidence_handler"),
-        Mock(pid=4, name="blocking"),
-        Mock(pid=5, name="exporting_alerts"),
-        Mock(pid=6, name="alert_summary"),
-        Mock(pid=7, name="LLM"),
-    ]
-
-    process_manager.main.db.get_pid_of = lambda x: {
-        "evidence_handler": 3,
-        "blocking": 4,
-        "exporting_alerts": 5,
-        "alert_summary": 6,
-        "LLM": 7,
-    }.get(x)
-    process_manager.main.args.blocking = blocking_enabled
-    process_manager.main.db.get_disabled_modules = lambda: (
-        ["exporting_alerts"] if exporting_alerts_disabled else []
-    )
-
-    to_kill_first, to_kill_last = process_manager.get_hitlist_in_order()
-
-    assert [p.pid for p in to_kill_first] == expected_kill_first
-    assert [p.pid for p in to_kill_last] == expected_kill_last
-
-
-@pytest.mark.parametrize(
-    "alive_statuses, expected_alive_count",
-    [  # Testcase1: two processes still alive
-        ([True, True, False], 2),
-        # Testcase2: all processes finished
-        ([False, False, False], 0),
-        # Tetscase3: first and third processes alive
-        ([True, False, True], 2),
-    ],
-)
-def test_wait_for_processes_to_finish(alive_statuses, expected_alive_count):
-    process_manager = ModuleFactory().create_process_manager_obj()
-
-    # create mock process objects based on the `alive_statuses`
-    mock_processes = [
-        Mock(name=f"Process{i}") for i in range(len(alive_statuses))
-    ]
-
-    # set up the is_alive of each process
-    for i, process in enumerate(mock_processes):
-        process.is_alive.return_value = alive_statuses[i]
-
-    with patch("time.time", side_effect=[0, 3.1]):
-        with patch.object(
-            process_manager, "print_stopped_module"
-        ) as mock_print_stopped:
-            alive_processes = process_manager.wait_for_processes_to_finish(
-                mock_processes
-            )
-
-    # assertions
-    # verify the number of alive processes matches the expected count
-    assert len(alive_processes) == expected_alive_count, (
-        f"Expected {expected_alive_count} alive processes, but got "
-        f"{len(alive_processes)}"
-    )
-
-    # verify the `print_stopped_module` method is called for all stopped processes
-    expected_stopped_count = len(alive_statuses) - expected_alive_count
-    assert mock_print_stopped.call_count == expected_stopped_count, (
-        f"Expected `print_stopped_module` to be called "
-        f"{expected_stopped_count} times, "
-        f"but it was called {mock_print_stopped.call_count} times"
-    )
-
-
-@pytest.mark.parametrize(
-    "end_date_str, start_time_str, expected_analysis_time",
-    [
-        # Test case 1: Analysis time is 10 minutes
-        (
-            1680343800,
-            1680343200,
-            10.0,
-        ),  # "2023-04-01 10:10:00", "2023-04-01 10:00:00"
-        # Test case 2: Analysis time is 1 hour
-        (
-            1680346800,
-            1680343200,
-            60.0,
-        ),  # "2023-04-01 11:00:00", "2023-04-01 10:00:00"
-        # Test case 3: Analysis time is less than a minute
-        (
-            1680343230,
-            1680343200,
-            0.5,
-        ),  # "2023-04-01 10:00:30", "2023-04-01 10:00:00"
-    ],
-)
-def test_get_analysis_time(
-    end_date_str, start_time_str, expected_analysis_time
-):
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.db.get_slips_start_time.return_value = start_time_str
-
-    with patch.object(utils, "convert_ts_format", return_value=end_date_str):
-        analysis_time = process_manager.get_analysis_time()
-
-    assert analysis_time == (expected_analysis_time, end_date_str)
-
-
-@pytest.mark.parametrize(
-    "message, msg_recvd_in_control_channel, expected_result",
-    [
-        # Test case 1: Message is None
-        (None, True, False),
-        # Test case 2: Message doesn't contain "stop_slips"
-        ({"data": "some_other_message"}, True, False),
-        # Test case 3: Wrapped plain-text messages should be decoded first
-        (
-            {
-                "data": json.dumps(
-                    {"text": "stop_slips", "version": "test-version"}
-                )
-            },
-            True,
-            True,
-        ),
-        # Test case 3: Message contains
-        # "stop_slips" but not intended for control channel
-        ({"data": "stop_slips"}, False, False),
-        # Test case 4: Message contains
-        # "stop_slips" and intended for control channel
-        ({"data": "stop_slips"}, True, True),
-    ],
-)
-def test_is_stop_msg_received(
-    message, msg_recvd_in_control_channel, expected_result
-):
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.c1.get_message.return_value = message
-
-    with patch(
-        "slips_files.common.slips_utils.utils.is_msg_intended_for"
-    ) as mock_is_intended_for:
-        mock_is_intended_for.return_value = msg_recvd_in_control_channel
-        assert process_manager.is_stop_msg_received() == expected_result
-
-
-@pytest.mark.parametrize(
-    "mock_return_value, expected_result",
-    [  # Testcase1: Debugger inactive
-        (None, False),
-        # Testcase2: Debugger active
-        (Mock(), True),
-    ],
-)
-def test_is_debugger_active(mock_return_value, expected_result):
-    mock_conf = Mock()
-    mock_conf.get_bootstrapping_setting.return_value = (False, [])
-    process_manager = ProcessManager(mock_conf)  # This line should now work
-    with patch("sys.gettrace", return_value=mock_return_value):
-        assert process_manager.is_debugger_active() == expected_result
-
-
-@pytest.mark.parametrize(
-    "debugger_active, input_type, is_interface, expected",
-    [
-        # Test case 1: Debugger active
-        (True, InputType.PCAP, False, True),
-        # Test case 2: Stdin input
-        (False, InputType.STDIN, True, True),
-        # Test case 3: Cyst input
-        (False, InputType.CYST, True, True),
-        # Test case 4: Interface input
-        (False, InputType.PCAP, True, True),
-        # Test case 5: Normal case (should stop)
-        (False, InputType.PCAP, False, False),
-    ],
-)
-def test_should_run_non_stop(
-    debugger_active, input_type, is_interface, expected
-):
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.is_debugger_active = Mock(return_value=debugger_active)
-    process_manager.main.input_type = input_type
-    process_manager.main.db.is_running_non_stop = Mock(
-        return_value=is_interface
-    )
-
-    assert process_manager.should_run_non_stop() == expected
-
-
-@pytest.mark.parametrize(
-    "input_acquired, profiler_acquired, expected_result",
-    [  # Test case 1: Both semaphores are not acquired
-        (False, False, False),
-        # Test case 2: Only input semaphore is acquired
-        (True, False, False),
-        # Testcase 3: Only profiler acquired
-        (False, True, False),
-        # Testcase 4: Both semaphores are acquired
-        (True, True, True),
-    ],
-)
-def test_is_done_receiving_new_flows(
-    input_acquired, profiler_acquired, expected_result
-):
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.can_acquire_semaphore = Mock(
-        side_effect=[input_acquired, profiler_acquired]
-    )
-    assert process_manager.is_done_receiving_new_flows() == expected_result
-
-
-def test_stop_llm_stack_if_llm_module_stopped_kills_dependents() -> None:
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.module_dependencies = {
-        Modules.LLM_PROXY: (),
-        Modules.REGEX_GENERATOR: (Modules.LLM_PROXY,),
-        Modules.ALERT_SUMMARY: (Modules.LLM_PROXY,),
-    }
-    process_manager.main.db.get_pid_of.side_effect = lambda module_name: {
-        Modules.LLM_PROXY: 100,
-        Modules.REGEX_GENERATOR: 101,
-        Modules.ALERT_SUMMARY: 102,
-    }.get(module_name)
-
-    with patch(
-        "managers.process_manager.shutdown_mixin.os.kill",
-        side_effect=ProcessLookupError,
-    ), patch.object(
-        process_manager, "kill_process_tree"
-    ) as mock_kill_process_tree, patch.object(
-        process_manager.main, "print"
-    ) as mock_print:
-        process_manager._stop_llm_stack_if_llm_module_stopped()
-
-    assert mock_kill_process_tree.call_args_list == [call(101), call(102)]
-    assert process_manager.stopped_modules == [
-        str(Modules.REGEX_GENERATOR),
-        str(Modules.ALERT_SUMMARY),
-        str(Modules.LLM_PROXY),
-    ]
-    mock_print.assert_called_once_with(
-        "Stopping modules because llm_proxy stopped: "
-        "['Modules.REGEX_GENERATOR', 'Modules.ALERT_SUMMARY']"
-    )
-
-
-@pytest.mark.parametrize(
-    "llm_enabled, os_kill_side_effect, stopped_modules",
-    [
-        (False, None, []),
-        (True, None, []),
-        (True, ProcessLookupError, [str(Modules.REGEX_GENERATOR)]),
-    ],
-)
-def test_stop_llm_stack_if_llm_module_stopped_skips_unneeded_shutdown(
-    llm_enabled: bool,
-    os_kill_side_effect: type[BaseException] | None,
-    stopped_modules: list[str],
-) -> None:
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.module_dependencies = {
-        Modules.LLM_PROXY: (),
-        Modules.REGEX_GENERATOR: (Modules.LLM_PROXY,),
-    }
-    process_manager.main.conf.llm_enabled.return_value = llm_enabled
-    process_manager.main.db.get_pid_of.side_effect = lambda module_name: {
-        Modules.LLM_PROXY: 100,
-        Modules.REGEX_GENERATOR: 101,
-    }.get(module_name)
-    process_manager.stopped_modules = stopped_modules
-
-    with patch(
-        "managers.process_manager.shutdown_mixin.os.kill",
-        side_effect=os_kill_side_effect,
-    ), patch.object(
-        process_manager, "kill_process_tree"
-    ) as mock_kill_process_tree, patch.object(
-        process_manager.main, "print"
-    ) as mock_print:
-        process_manager._stop_llm_stack_if_llm_module_stopped()
-
-    if llm_enabled and os_kill_side_effect is ProcessLookupError:
-        mock_kill_process_tree.assert_called_once_with(101)
-        mock_print.assert_called_once_with(
-            "Stopping modules because llm_proxy stopped: "
-            "['Modules.REGEX_GENERATOR']"
-        )
-    else:
-        mock_kill_process_tree.assert_not_called()
-        mock_print.assert_not_called()
-
-
-@pytest.mark.parametrize(
-    "live_update, stop_received, done_receiving, expected_result",
-    [
-        (True, False, False, True),
-        (False, True, False, True),
-        (False, False, True, True),
-        (False, False, False, False),
-    ],
-)
-def test_should_stop_slips(
-    live_update, stop_received, done_receiving, expected_result
-):
-    """
-    Test whether Slips should stop for live updates, stop messages, or done input.
-
-    Parameters:
-    live_update: Whether a live update is in progress.
-    stop_received: Whether a stop message was received.
-    done_receiving: Whether input and profiler finished processing.
-    expected_result: Expected stop decision.
-
-    Return:
-    None.
-    """
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.is_slips_live_updating_event.is_set = Mock(
-        return_value=live_update
-    )
-    process_manager.is_stop_msg_received = Mock(return_value=stop_received)
-    process_manager.is_done_receiving_new_flows = Mock(
-        return_value=done_receiving
-    )
-    process_manager._did_a_core_module_fail = Mock(return_value=False)
-    process_manager.all_children_started = True
-
-    assert process_manager.should_stop_slips() == expected_result
-
-
-@pytest.mark.parametrize(
-    "input_exitcode, profiler_exitcode, evidence_exitcode, expected_result",
-    [
-        (None, None, None, False),
-        (0, None, None, False),
-        (None, 1, None, True),
-        (0, 1, None, False),
-        (None, None, 0, True),
-    ],
-)
-def test_did_a_core_module_fail_for_file_input(
-    input_exitcode: int | None,
-    profiler_exitcode: int | None,
-    evidence_exitcode: int | None,
-    expected_result: bool,
-) -> None:
-    """
-    Test core module failure detection for finite file inputs.
-
-    Parameters:
-    input_exitcode: Input process exit code.
-    profiler_exitcode: Profiler process exit code.
-    evidence_exitcode: Evidence process exit code.
-    expected_result: Expected failure detection result.
-
-    Return:
-    None.
-    """
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.input_process = Mock()
-    process_manager.profiler_process = Mock()
-    process_manager.evidence_process = Mock()
-    process_manager.input_process.exitcode = input_exitcode
-    process_manager.profiler_process.exitcode = profiler_exitcode
-    process_manager.evidence_process.exitcode = evidence_exitcode
-    process_manager.main.db.is_running_non_stop.return_value = False
-
-    assert process_manager._did_a_core_module_fail() == expected_result
-
-
-@pytest.mark.parametrize(
-    "input_type",
-    [
-        InputType.INTERFACE,
-        InputType.STDIN,
-        InputType.CYST,
-    ],
-)
-@pytest.mark.parametrize(
-    "input_exitcode, profiler_exitcode, evidence_exitcode, expected_result",
-    [
-        (None, None, None, False),
-        (0, None, None, True),
-        (None, 1, None, True),
-        (None, None, 0, True),
-    ],
-)
-def test_did_a_core_module_fail_for_forever_growing_input(
-    input_type: InputType,
-    input_exitcode: int | None,
-    profiler_exitcode: int | None,
-    evidence_exitcode: int | None,
-    expected_result: bool,
-) -> None:
-    """
-    Test core module failure detection for forever-growing inputs.
-
-    Parameters:
-    input_exitcode: Input process exit code.
-    profiler_exitcode: Profiler process exit code.
-    evidence_exitcode: Evidence process exit code.
-    expected_result: Expected failure detection result.
-
-    Return:
-    None.
-    """
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.input_type = input_type
-    process_manager.input_process = Mock()
-    process_manager.profiler_process = Mock()
-    process_manager.evidence_process = Mock()
-    process_manager.input_process.exitcode = input_exitcode
-    process_manager.profiler_process.exitcode = profiler_exitcode
-    process_manager.evidence_process.exitcode = evidence_exitcode
-    process_manager.main.db.is_running_non_stop.return_value = True
-
-    assert process_manager._did_a_core_module_fail() == expected_result
-
-
-def test_should_stop_slips_sets_core_module_failure() -> None:
-    """
-    Test should_stop_slips marks core module failures for shutdown handling.
-
-    Return:
-    None.
-    """
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.is_slips_live_updating_event.is_set = Mock(
-        return_value=False
-    )
-    process_manager._did_a_core_module_fail = Mock(return_value=True)
-    process_manager.is_stop_msg_received = Mock()
-    process_manager.is_done_receiving_new_flows = Mock()
-    process_manager.all_children_started = True
-
-    assert process_manager.should_stop_slips() is True
-    assert process_manager.core_module_failure is True
-    process_manager.is_stop_msg_received.assert_not_called()
-    process_manager.is_done_receiving_new_flows.assert_not_called()
-
-
-def test_shutdown_gracefully_handles_core_module_failure() -> None:
-    """
-    Test shutdown avoids waiting for modules after a core module failure.
-
-    Return:
-    None.
-    """
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.core_module_failure = True
-    process_manager.main.args.stopdaemon = False
-    process_manager.main.args.save = False
-    process_manager.main.input_information = "test_input"
-    process_manager.main.conf.wait_for_modules_to_finish.return_value = 1
-    process_manager.main.conf.export_labeled_flows.return_value = False
-    process_manager.main.db.get_flows_count.return_value = 42
-    process_manager.main.db.check_tw_to_close = Mock()
-    process_manager.main.db.close_all_dbs = Mock()
-    process_manager.main.metadata_man = Mock()
-    process_manager.main.profilers_manager = Mock()
-    process_manager.main.store_zeek_dir_copy = Mock()
-    process_manager.main.delete_zeek_files = Mock()
-    process_manager.get_hitlist_in_order = Mock(return_value=([], []))
-    process_manager.shutdown_interactive = Mock()
-    process_manager.kill_all_children = Mock()
-    process_manager.get_analysis_time = Mock(
-        return_value=(1.23, "2026/05/21 12:00:00")
-    )
-    process_manager.is_slips_live_updating_event.is_set = Mock(
-        return_value=False
-    )
-
-    with patch(
-        "managers.process_manager.shutdown_mixin.multiprocessing.active_children",
-        return_value=[],
-    ):
-        process_manager.shutdown_gracefully()
-
-    process_manager.shutdown_interactive.assert_not_called()
-    assert process_manager.kill_all_children.call_count == 2
-    process_manager.main.print.assert_any_call(
-        "[Process Manager] Slips didn't shutdown gracefully - "
-        "Core module failure.\n",
-        log_to_logfiles_only=True,
-    )
-
-
-@pytest.mark.parametrize(
-    "mode, expected_print_function",
-    [  # Test case 1: Daemonized mode
-        ("daemonized", "main.daemon.print"),
-        # Test case 2: Normal mode
-        ("interactive", "main.print"),
-    ],
-)
-def test_get_print_function(mode, expected_print_function):
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.mode = mode
-
-    process_manager.main.daemon = Mock()
-    process_manager.main.daemon.print = Mock()
-    process_manager.main.print = Mock()
-
-    print_function = process_manager.get_print_function()
-
-    expected_function = eval(f"process_manager.{expected_print_function}")
-
-    assert print_function == expected_function
-
-    print_function()
-    expected_function.assert_called_once()
-
-
-def test_print_stopped_module():
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.children = [Mock(), Mock()]
-    process_manager.stopped_modules = []
-
-    with patch(
-        "managers.process_manager.reporting_mixin.green",
-        side_effect=["green_module_name", "green_count"],
-    ), patch.object(process_manager.main, "print") as mock_print:
-        process_manager.print_stopped_module("TestModule")
-
-        assert "TestModule" in process_manager.stopped_modules
-        mock_print.assert_called_once()
-
-        printed_str = mock_print.call_args[0][0]
-        assert "green_module_name" in printed_str
-        assert "Stopped" in printed_str
-        assert "green_count left" in printed_str
-
-        process_manager.print_stopped_module("testmodule")
-
-        assert mock_print.call_count == 1
-
-
-def test_kill_daemon_children_excludes_thread_pids_from_logging_count():
-    """Exclude stored thread PIDs from daemon-child shutdown log counts."""
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.db.get_pids.return_value = {
-        "module_one": 123,
-        "module_thread": 456,
-        "module_two": 789,
-    }
-
-    with patch.object(
-        process_manager, "kill_process_tree"
-    ) as mock_kill_process_tree, patch.object(
-        process_manager, "print_stopped_module"
-    ) as mock_print_stopped_module:
-        process_manager.kill_daemon_children()
-
-    assert mock_kill_process_tree.call_args_list == [call(123), call(789)]
-    assert mock_print_stopped_module.call_args_list == [
-        call("module_one", total_modules=2),
-        call("module_two", total_modules=2),
-    ]
-
-
-def test_kill_process_tree_kills_descendants_before_parent():
-    """Kill descendants before their parent can reparent them."""
-    process_manager = ModuleFactory().create_process_manager_obj()
-    parent_children = Mock()
-    parent_children.read.return_value = "456\n"
-    child_children = Mock()
-    child_children.read.return_value = ""
-
-    with patch(
-        "managers.process_manager.shutdown_mixin.os.popen",
-        side_effect=[parent_children, child_children],
-    ), patch("managers.process_manager.shutdown_mixin.os.kill") as mock_kill:
-        process_manager.kill_process_tree(123)
-
-    assert mock_kill.call_args_list == [
-        call(456, signal.SIGKILL),
-        call(123, signal.SIGKILL),
-    ]
-
-
-def test_shutdown_interactive_signals_evidence_handler_after_other_modules_stop():
-    """Delay the evidence-handler shutdown signal until earlier modules stop."""
-    process_manager = ModuleFactory().create_process_manager_obj()
-    first_process = Mock()
-    last_process = Mock()
-
-    with patch.object(
-        process_manager,
-        "wait_for_processes_to_finish",
-        side_effect=[[], []],
-    ) as mock_wait, patch.object(
-        process_manager.evidence_handler_termination_event, "set"
-    ) as mock_set:
-        result = process_manager.shutdown_interactive(
-            [first_process], [last_process]
-        )
-
-    assert result == (None, None)
-    assert mock_wait.call_args_list == [
-        call([first_process]),
-        call([last_process]),
-    ]
-    mock_set.assert_called_once_with()
-
-
-def test_shutdown_interactive_does_not_signal_evidence_handler_while_modules_are_pending():
-    """Keep the evidence handler running while earlier modules are pending."""
-    process_manager = ModuleFactory().create_process_manager_obj()
-    pending_process = Mock()
-    last_process = Mock()
-
-    with patch.object(
-        process_manager,
-        "wait_for_processes_to_finish",
-        return_value=[pending_process],
-    ) as mock_wait, patch.object(
-        process_manager, "warn_about_pending_modules"
-    ) as mock_warn, patch.object(
-        process_manager.evidence_handler_termination_event, "set"
-    ) as mock_set:
-        result = process_manager.shutdown_interactive(
-            [pending_process], [last_process]
-        )
-
-    assert result == ([pending_process], [last_process])
-    mock_wait.assert_called_once_with([pending_process])
-    mock_warn.assert_called_once_with([pending_process, last_process])
-    mock_set.assert_not_called()
-
-
-def test_start_profiler_process():
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.bloom_filters_man = Mock()
-    with patch(
-        "managers.process_manager.startup_mixin.Profiler"
-    ) as mock_profiler, patch.object(
-        process_manager.is_profiler_done_starting_initial_workers_event,
-        "wait",
-    ):
-        mock_profiler_process = Mock()
-        mock_profiler.return_value = mock_profiler_process
-        mock_profiler_process.pid = 67890
-
-        result = process_manager.start_profiler_process()
-
-        assert result == mock_profiler_process
-        mock_profiler.assert_called_once_with(
-            process_manager.main.logger,
-            process_manager.main.args.output,
-            process_manager.main.redis_port,
-            process_manager.termination_event,
-            process_manager.main.args,
-            process_manager.main.conf,
-            process_manager.main.pid,
-            process_manager.main.bloom_filters_man,
-            is_profiler_done_semaphore=(
-                process_manager.is_profiler_done_semaphore
-            ),
-            profiler_queue=process_manager.profiler_queue,
-            is_profiler_done_event=process_manager.is_profiler_done_event,
-            is_input_done_event=process_manager.is_input_done_event,
-            is_input_failed_event=process_manager.is_input_failed_event,
-            is_profiler_done_starting_initial_workers_event=(
-                process_manager.is_profiler_done_starting_initial_workers_event
-            ),
-        )
-        mock_profiler_process.start.assert_called_once()
-        process_manager.main.print.assert_called_once()
-        process_manager.main.db.store_pid.assert_called_once_with(
-            "Profiler", 67890
-        )
-
-
-@pytest.mark.parametrize(
-    "input_type, should_wait",
-    [
-        (InputType.INTERFACE, True),
-        (InputType.PCAP, False),
-    ],
-)
-def test_start_profiler_process_waits_for_initial_workers(
-    input_type, should_wait
-):
-    """
-    Test that only interface profiler startup waits for initial workers.
-
-    Parameters:
-    input_type: Input type used for the run.
-    should_wait: Whether profiler startup should wait for initial workers.
-
-    Return:
-    None.
-    """
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.bloom_filters_man = Mock()
-    process_manager.main.input_type = input_type
-    wait_event = (
-        process_manager.is_profiler_done_starting_initial_workers_event
-    )
-
-    with patch(
-        "managers.process_manager.startup_mixin.Profiler"
-    ) as mock_profiler, patch.object(wait_event, "wait") as mock_wait:
-        mock_profiler_process = Mock()
-        mock_profiler.return_value = mock_profiler_process
-        mock_profiler_process.pid = 67890
-
-        process_manager.start_profiler_process()
-
-    if should_wait:
-        mock_wait.assert_called_once_with(30)
-    else:
-        mock_wait.assert_not_called()
-
-
-@pytest.mark.parametrize(
-    "output_dir, redis_port",
-    [
-        # Test case 1: Default output directory and Redis port
-        ("output", 6379),
-        # Test case 2: Custom output directory and Redis port
-        ("/custom/output", 6380),
-    ],
-)
-def test_start_evidence_process(output_dir, redis_port):
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.bloom_filters_man = Mock()
-    process_manager.main.args.output = output_dir
-    process_manager.main.redis_port = redis_port
-
-    with patch(
-        "managers.process_manager.startup_mixin.EvidenceHandler"
-    ) as mock_evidence:
-        mock_evidence_process = Mock()
-        mock_evidence.return_value = mock_evidence_process
-        mock_evidence_process.pid = 13579
-
-        result = process_manager.start_evidence_process()
-
-        assert result == mock_evidence_process
-        mock_evidence.assert_called_once_with(
-            process_manager.main.logger,
-            output_dir,
-            redis_port,
-            process_manager.evidence_handler_termination_event,
-            process_manager.main.args,
-            process_manager.main.conf,
-            process_manager.main.pid,
-            process_manager.main.bloom_filters_man,
-        )
-        mock_evidence_process.start.assert_called_once()
-        process_manager.main.print.assert_called_once()
-        process_manager.main.db.store_pid.assert_called_once_with(
-            "evidence_handler", 13579
-        )
-
-
-def test_print_started_module():
-    process_manager = ModuleFactory().create_process_manager_obj()
-    with patch(
-        "managers.process_manager.reporting_mixin.green",
-        return_value="green_module_name",
-    ), patch.object(process_manager.main, "print") as mock_print:
-        process_manager.print_started_module(
-            Modules.CESNET, 12345, "Test description"
-        )
-
-        mock_print.assert_called_once_with(
-            "\t\tStarting green_module_name module "
-            "(Test description) [PID green_module_name]",
-            1,
-            0,
-        )
-
-
-@pytest.mark.parametrize(
-    "local_files, ti_feeds, ports_called, orgs_called, "
-    "whitelist_called, print_called, asyncio_called",
-    [  # Testcase1: Update both
-        (True, True, True, True, True, True, True),
-        # Testcase2: Update local only
-        (True, False, True, True, True, False, False),
-        # Testcase3: Update TI only
-        (False, True, False, False, False, True, True),
-        # Testcase4: Don't update
-        (False, False, False, False, False, False, False),
-    ],
-)
-@patch("asyncio.run")
-@patch("managers.process_manager.startup_mixin.Lock")
-def test_start_update_manager(
-    mock_lock,
-    mock_asyncio_run,
-    local_files,
-    ti_feeds,
-    ports_called,
-    orgs_called,
-    whitelist_called,
-    print_called,
-    asyncio_called,
-):
-    process_manager = ModuleFactory().create_process_manager_obj()
-    process_manager.main.args.output = "output"
-    mock_lock_instance = Mock()
-    mock_lock.return_value.__enter__.return_value = mock_lock_instance
-
-    mock_update_manager = Mock()
-    with patch(
-        "managers.process_manager.startup_mixin.FeedsUpdateManager",
-        return_value=mock_update_manager,
-    ):
-        process_manager.start_update_manager(
-            local_files=local_files, ti_feeds=ti_feeds
-        )
-
-    assert mock_update_manager.update_ports_info.called is ports_called
-    assert mock_update_manager.update_org_files.called is orgs_called
-    assert (
-        mock_update_manager.update_local_whitelist.called is whitelist_called
-    )
-    assert mock_update_manager.print.called is print_called
-    assert mock_asyncio_run.called is asyncio_called
